+++ packages/assets/src/wallets/im-token.tsx
@@ -1,4 +1,4 @@
-import type { WalletMetadata } from '@ant-design/web3-common';
+import { ChainType, type WalletMetadata } from '@ant-design/web3-common';
 import { ImTokenColorful } from '@ant-design/web3-icons';
 
 export const metadata_imToken: WalletMetadata = {
@@ -8,4 +8,13 @@ export const metadata_imToken: WalletMetadata = {
   app: {
     link: 'https://token.im/download',
   },
+  transferQRCodeFormatter: (params) => {
+    const { toAddress, chainId, amount, tokenAddress = '', decimal } = params;
+    if (tokenAddress) {
+      return `ethereum:${toAddress}@${chainId}?contractAddress=${tokenAddress}&decimal=${decimal}&value=${amount}`;
+    } else {
+      return `ethereum:${toAddress}@${chainId}?decimal=18&value=${amount}`;
+    }
+  },
+  supportChainTypes: [ChainType.EVM],
 };