+++ packages/ethers-v5/src/wallets.ts
@@ -1,1 +1,2 @@
+/* v8 ignore start */
 export * from '@ant-design/web3-ethers/wallets';