+++ packages/wagmi/src/wagmi-provider/__tests__/nft.test.tsx
@@ -50,6 +50,7 @@ vi.mock('wagmi', () => {
       return {};
     },
     useEnsName: () => ({}),
+    useEnsAvatar: () => ({}),
   };
 });