+++ packages/wagmi/src/wagmi-provider/__tests__/switch-chain.test.tsx
@@ -70,6 +70,7 @@ vi.mock('wagmi', () => {
       };
     },
     useEnsName: () => ({}),
+    useEnsAvatar: () => ({}),
   };
 });