+++ packages/wagmi/src/wagmi-provider/config-provider.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  fillAddressWith0x,
   Web3ConfigProvider,
   type Account,
   type Chain,
@@ -13,6 +12,7 @@ import {
   useBalance,
   useConfig,
   useConnect,
+  useEnsAvatar,
   useEnsName,
   useSwitchChain,
   type Connector as WagmiConnector,
@@ -43,35 +43,26 @@ export const AntDesignWeb3ConfigProvider: React.FC<AntDesignWeb3ConfigProviderPr
     walletFactorys,
     availableChains,
     availableConnectors,
-    ens,
+    ens = true,
     balance,
     locale,
     eip6963,
   } = props;
   const { address, isDisconnected, chain } = useAccount();
   const config = useConfig();
-  const [account, setAccount] = React.useState<Account | undefined>();
   const { connectAsync } = useConnect();
   const { switchChain } = useSwitchChain();
-  const { data: balanceData } = useBalance({
-    address: balance && account ? fillAddressWith0x(account.address) : undefined,
-  });
+  const { data: balanceData } = useBalance({ address });
   const { data: ensName } = useEnsName({ address });
-
-  React.useEffect(() => {
-    if (!address || isDisconnected) {
-      setAccount(undefined);
-    } else if (ens && ensName) {
-      setAccount({
-        address: fillAddressWith0x(address),
-        name: ensName as string,
-      });
-    } else {
-      setAccount({
-        address: fillAddressWith0x(address),
-      });
-    }
-  }, [address, isDisconnected, ens, ensName]);
+  const { data: ensAvatar } = useEnsAvatar({ name: ensName ?? undefined });
+  const account: Account | undefined =
+    address && !isDisconnected
+      ? {
+          address,
+          name: ensName && ens ? ensName : undefined,
+          avatar: ensAvatar ?? undefined,
+        }
+      : undefined;
 
   const findConnectorByName = (name: string): WagmiConnector | undefined => {
     const commonConnector = availableConnectors.find(